[SRE-1722] plugin version "1.0" from grafana.internal
Summary
=======

I found that the rockset-grafana-backend plugin we have installed on
grafana.internal.rockset.io doesn't match with any commit or branch in
source control. The first clue is that on grafana.internal.rockset.io
the plugin is version 1.0, and in source control this version never
changes from 0.1.0.

This PR is simply an rsync of that plugin dir with an exclusion for .git
and .gitignore.

I don't think we should ship it in this state, there appear to be some
directories to nowhere and some things that I don't think we even use
anymore, like .circleci.

Test Plan
=========

TBD