Ensure that we are dereferencing elements that exists
Summary
We were derefencing variables that could be nil.
For instance if the query don't return an error but just no lines then
doing err.Error() was causing a nil/null derefencing.
Similarly if the field array in the frame was has only one element doing
```
frame.Fields[1].Len()
 ```

Would result in a crash because we would derefence a nil pointer.

We are fixing this here. Also we are returning a nicer message when the
query results are empty.

Testing

Used a query that is returning no line, before the change this would
crash the plug-in after the change it's printing a nice(r) message
about the fact that the query is not returning lines.